replace int by size_t

+++ src/RcppExports.cpp
@@ -19,16 +19,16 @@ BEGIN_RCPP
 END_RCPP
 }
 // singletree_2
-Rcpp::List singletree_2(arma::vec y, arma::mat X, int depth, int max_depth, int Nmin, double tau, double sigma, double alpha, double beta);
+Rcpp::List singletree_2(arma::vec y, arma::mat X, size_t depth, size_t max_depth, size_t Nmin, double tau, double sigma, double alpha, double beta);
 RcppExport SEXP _SpikySmoothTrees_singletree_2(SEXP ySEXP, SEXP XSEXP, SEXP depthSEXP, SEXP max_depthSEXP, SEXP NminSEXP, SEXP tauSEXP, SEXP sigmaSEXP, SEXP alphaSEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< arma::vec >::type y(ySEXP);
     Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
-    Rcpp::traits::input_parameter< int >::type depth(depthSEXP);
-    Rcpp::traits::input_parameter< int >::type max_depth(max_depthSEXP);
-    Rcpp::traits::input_parameter< int >::type Nmin(NminSEXP);
+    Rcpp::traits::input_parameter< size_t >::type depth(depthSEXP);
+    Rcpp::traits::input_parameter< size_t >::type max_depth(max_depthSEXP);
+    Rcpp::traits::input_parameter< size_t >::type Nmin(NminSEXP);
     Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
     Rcpp::traits::input_parameter< double >::type sigma(sigmaSEXP);
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
@@ -38,16 +38,16 @@ BEGIN_RCPP
 END_RCPP
 }
 // singletree_3
-Rcpp::List singletree_3(arma::vec y, arma::mat X, int depth, int max_depth, int Nmin, double tau, double sigma, double alpha, double beta);
+Rcpp::List singletree_3(arma::vec y, arma::mat X, size_t depth, size_t max_depth, size_t Nmin, double tau, double sigma, double alpha, double beta);
 RcppExport SEXP _SpikySmoothTrees_singletree_3(SEXP ySEXP, SEXP XSEXP, SEXP depthSEXP, SEXP max_depthSEXP, SEXP NminSEXP, SEXP tauSEXP, SEXP sigmaSEXP, SEXP alphaSEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< arma::vec >::type y(ySEXP);
     Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
-    Rcpp::traits::input_parameter< int >::type depth(depthSEXP);
-    Rcpp::traits::input_parameter< int >::type max_depth(max_depthSEXP);
-    Rcpp::traits::input_parameter< int >::type Nmin(NminSEXP);
+    Rcpp::traits::input_parameter< size_t >::type depth(depthSEXP);
+    Rcpp::traits::input_parameter< size_t >::type max_depth(max_depthSEXP);
+    Rcpp::traits::input_parameter< size_t >::type Nmin(NminSEXP);
     Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
     Rcpp::traits::input_parameter< double >::type sigma(sigmaSEXP);
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
@@ -57,16 +57,16 @@ BEGIN_RCPP
 END_RCPP
 }
 // singletree
-Rcpp::List singletree(arma::vec y, arma::mat X, int depth, int max_depth, int Nmin, double tau, double sigma, double alpha, double beta);
+Rcpp::List singletree(arma::vec y, arma::mat X, size_t depth, size_t max_depth, size_t Nmin, double tau, double sigma, double alpha, double beta);
 RcppExport SEXP _SpikySmoothTrees_singletree(SEXP ySEXP, SEXP XSEXP, SEXP depthSEXP, SEXP max_depthSEXP, SEXP NminSEXP, SEXP tauSEXP, SEXP sigmaSEXP, SEXP alphaSEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< arma::vec >::type y(ySEXP);
     Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
-    Rcpp::traits::input_parameter< int >::type depth(depthSEXP);
-    Rcpp::traits::input_parameter< int >::type max_depth(max_depthSEXP);
-    Rcpp::traits::input_parameter< int >::type Nmin(NminSEXP);
+    Rcpp::traits::input_parameter< size_t >::type depth(depthSEXP);
+    Rcpp::traits::input_parameter< size_t >::type max_depth(max_depthSEXP);
+    Rcpp::traits::input_parameter< size_t >::type Nmin(NminSEXP);
     Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
     Rcpp::traits::input_parameter< double >::type sigma(sigmaSEXP);
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
@@ -76,7 +76,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // train_forest_2
-Rcpp::List train_forest_2(arma::mat y, arma::mat X, arma::mat Xtest, int M, int L, int N_sweeps, arma::vec max_depth, int Nmin, double alpha, double beta, double tau, bool draw_sigma, bool verbose);
+Rcpp::List train_forest_2(arma::mat y, arma::mat X, arma::mat Xtest, size_t M, size_t L, size_t N_sweeps, arma::vec max_depth, size_t Nmin, double alpha, double beta, double tau, bool draw_sigma, bool verbose);
 RcppExport SEXP _SpikySmoothTrees_train_forest_2(SEXP ySEXP, SEXP XSEXP, SEXP XtestSEXP, SEXP MSEXP, SEXP LSEXP, SEXP N_sweepsSEXP, SEXP max_depthSEXP, SEXP NminSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP tauSEXP, SEXP draw_sigmaSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -84,11 +84,11 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::mat >::type y(ySEXP);
     Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
     Rcpp::traits::input_parameter< arma::mat >::type Xtest(XtestSEXP);
-    Rcpp::traits::input_parameter< int >::type M(MSEXP);
-    Rcpp::traits::input_parameter< int >::type L(LSEXP);
-    Rcpp::traits::input_parameter< int >::type N_sweeps(N_sweepsSEXP);
+    Rcpp::traits::input_parameter< size_t >::type M(MSEXP);
+    Rcpp::traits::input_parameter< size_t >::type L(LSEXP);
+    Rcpp::traits::input_parameter< size_t >::type N_sweeps(N_sweepsSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type max_depth(max_depthSEXP);
-    Rcpp::traits::input_parameter< int >::type Nmin(NminSEXP);
+    Rcpp::traits::input_parameter< size_t >::type Nmin(NminSEXP);
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
     Rcpp::traits::input_parameter< double >::type beta(betaSEXP);
     Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
@@ -99,7 +99,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // train_forest
-Rcpp::List train_forest(arma::mat y, arma::mat X, arma::mat Xtest, int M, int L, int N_sweeps, arma::vec max_depth, int Nmin, double alpha, double beta, double tau, bool draw_sigma);
+Rcpp::List train_forest(arma::mat y, arma::mat X, arma::mat Xtest, size_t M, size_t L, size_t N_sweeps, arma::vec max_depth, size_t Nmin, double alpha, double beta, double tau, bool draw_sigma);
 RcppExport SEXP _SpikySmoothTrees_train_forest(SEXP ySEXP, SEXP XSEXP, SEXP XtestSEXP, SEXP MSEXP, SEXP LSEXP, SEXP N_sweepsSEXP, SEXP max_depthSEXP, SEXP NminSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP tauSEXP, SEXP draw_sigmaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -107,11 +107,11 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::mat >::type y(ySEXP);
     Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
     Rcpp::traits::input_parameter< arma::mat >::type Xtest(XtestSEXP);
-    Rcpp::traits::input_parameter< int >::type M(MSEXP);
-    Rcpp::traits::input_parameter< int >::type L(LSEXP);
-    Rcpp::traits::input_parameter< int >::type N_sweeps(N_sweepsSEXP);
+    Rcpp::traits::input_parameter< size_t >::type M(MSEXP);
+    Rcpp::traits::input_parameter< size_t >::type L(LSEXP);
+    Rcpp::traits::input_parameter< size_t >::type N_sweeps(N_sweepsSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type max_depth(max_depthSEXP);
-    Rcpp::traits::input_parameter< int >::type Nmin(NminSEXP);
+    Rcpp::traits::input_parameter< size_t >::type Nmin(NminSEXP);
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
     Rcpp::traits::input_parameter< double >::type beta(betaSEXP);
     Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
