+++ src/randomkit.h
@@ -66,8 +66,8 @@
 typedef struct rk_state_
 {
 	unsigned long key[RK_STATE_LEN];
-	int pos;
-	int has_gauss; /* !=0: gauss contains a gaussian deviate */
+	size_t pos;
+	size_t has_gauss; /* !=0: gauss contains a gaussian deviate */
 	double gauss;
 }
 rk_state;
@@ -141,7 +141,7 @@ extern void rk_fill(void *buffer, size_t size, rk_state *state);
  * Warning: on most unixes RK_DEV_RANDOM will wait for enough entropy to answer
  * which can take a very long time on quiet systems.
  */
-extern rk_error rk_devfill(void *buffer, size_t size, int strong);
+extern rk_error rk_devfill(void *buffer, size_t size, size_t strong);
 
 /*
  * fill the buffer using rk_devfill if the random device is available and using
@@ -149,7 +149,7 @@ extern rk_error rk_devfill(void *buffer, size_t size, int strong);
  * parameters have the same meaning as rk_fill and rk_devfill
  * Returns RK_ENODEV if the device is unavailable, or RK_NOERR if it is
  */
-extern rk_error rk_altfill(void *buffer, size_t size, int strong,
+extern rk_error rk_altfill(void *buffer, size_t size, size_t strong,
 	rk_state *state);
 
 /*
